Fix namespaces after move to Error directory
