+++ app/Error/SentryCliServiceProvider.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\SearchService;
+namespace CultuurNet\UDB3\SearchService\Error;
 
 use CultuurNet\UDB3\SearchService\Error\SentryExceptionHandler;
+use CultuurNet\UDB3\SearchService\BaseServiceProvider;
 use Sentry\State\HubInterface;
 
 class SentryCliServiceProvider extends BaseServiceProvider
