+++ app/Error/SentryWebServiceProvider.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\SearchService;
+namespace CultuurNet\UDB3\SearchService\Error;
 
 use CultuurNet\UDB3\ApiGuard\ApiKey\ApiKey;
 use CultuurNet\UDB3\SearchService\Error\SentryExceptionHandler;
+use CultuurNet\UDB3\SearchService\BaseServiceProvider;
 use Sentry\State\HubInterface;
 
 class SentryWebServiceProvider extends BaseServiceProvider
