+++ app/Factory/ContainerFactory.php
@@ -14,9 +14,9 @@ use CultuurNet\UDB3\SearchService\Offer\OfferServiceProvider;
 use CultuurNet\UDB3\SearchService\Organizer\OrganizerServiceProvider;
 use CultuurNet\UDB3\SearchService\Place\PlaceServiceProvider;
 use CultuurNet\UDB3\SearchService\RoutingServiceProvider;
-use CultuurNet\UDB3\SearchService\SentryCliServiceProvider;
-use CultuurNet\UDB3\SearchService\SentryHubServiceProvider;
-use CultuurNet\UDB3\SearchService\SentryWebServiceProvider;
+use CultuurNet\UDB3\SearchService\Error\SentryCliServiceProvider;
+use CultuurNet\UDB3\SearchService\Error\SentryHubServiceProvider;
+use CultuurNet\UDB3\SearchService\Error\SentryWebServiceProvider;
 use League\Container\Container;
 use League\Container\ReflectionContainer;
 use Noodlehaus\Config;
